requirements.txt
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
fpdf==1.7.2
newsapi-python==0.2.7
beautifulsoup4==4.12.2
google-search-results==2.4.2
requests==2.31.0
python-dotenv==1.0.0
textstat==0.7.3
rich==13.7.0
typer==0.9.0
loguru==0.7.2
reportlab==4.0.8
python-docx==1.0.0
streamlit==1.29.0